requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.2.2
python-dotenv==1.0.0
     

